Whole codebase compiling
