+++ nodes/nomos-executor/executor/src/api/backend.rs
@@ -22,6 +22,7 @@ use nomos_api::{
     },
     Backend,
 };
+use nomos_blend_service::{ProofsGenerator, ProofsVerifier};
 use nomos_core::{
     da::{
         blob::{info::DispersedBlobInfo, metadata, LightShare, Share},
@@ -94,6 +95,8 @@ pub struct AxumBackend<
     TimeBackend,
     ApiAdapter,
     HttpStorageAdapter,
+    BlendProofsGenerator,
+    BlendProofsVerifier,
     const SIZE: usize,
 > {
     settings: AxumBackendSettings,
@@ -121,6 +124,8 @@ pub struct AxumBackend<
         TimeBackend,
         ApiAdapter,
         HttpStorageAdapter,
+        BlendProofsGenerator,
+        BlendProofsVerifier,
     )>,
 }
 
@@ -160,6 +165,8 @@ impl<
         TimeBackend,
         ApiAdapter,
         StorageAdapter,
+        BlendProofsGenerator,
+        BlendProofsVerifier,
         const SIZE: usize,
         RuntimeServiceId,
     > Backend<RuntimeServiceId>
@@ -185,6 +192,8 @@ impl<
         TimeBackend,
         ApiAdapter,
         StorageAdapter,
+        BlendProofsGenerator,
+        BlendProofsVerifier,
         SIZE,
     >
 where
@@ -292,6 +301,8 @@ where
     TimeBackend::Settings: Clone + Send + Sync,
     ApiAdapter: nomos_da_network_service::api::ApiAdapter + Send + Sync + 'static,
     StorageAdapter: storage::StorageAdapter<RuntimeServiceId> + Send + Sync + 'static,
+    BlendProofsGenerator: ProofsGenerator + Send + 'static,
+    BlendProofsVerifier: ProofsVerifier + Clone + Send + 'static,
     RuntimeServiceId: Debug
         + Sync
         + Send
@@ -305,6 +316,8 @@ where
                 SamplingNetworkAdapter,
                 SamplingStorage,
                 TimeBackend,
+                BlendProofsGenerator,
+                BlendProofsVerifier,
                 RuntimeServiceId,
                 SIZE,
             >,
@@ -381,7 +394,7 @@ where
         wait_until_services_are_ready!(
             &overwatch_handle,
             Some(Duration::from_secs(60)),
-            Cryptarchia<_, _, _, _, _, _, SIZE>,
+            Cryptarchia<_, _, _, _, _, _, _, _, SIZE>,
             DaVerifier<_, _, _, _, _, _>,
             nomos_da_network_service::NetworkService<_, _, _,_, _, _>,
             nomos_network::NetworkService<_, _>,
@@ -431,6 +444,8 @@ where
                         SamplingNetworkAdapter,
                         SamplingStorage,
                         TimeBackend,
+                        BlendProofsGenerator,
+                        BlendProofsVerifier,
                         RuntimeServiceId,
                         SIZE,
                     >,
@@ -445,6 +460,8 @@ where
                         SamplingNetworkAdapter,
                         SamplingStorage,
                         TimeBackend,
+                        BlendProofsGenerator,
+                        BlendProofsVerifier,
                         RuntimeServiceId,
                         SIZE,
                     >,
