+++ nodes/nomos-node/node/src/lib.rs
@@ -2,24 +2,10 @@ pub mod api;
 pub mod config;
 pub mod generic_services;
 
-use core::convert::Infallible;
-
-use async_trait::async_trait;
 use color_eyre::eyre::Result;
 use generic_services::VerifierMempoolAdapter;
 use kzgrs_backend::common::share::DaShare;
 pub use kzgrs_backend::dispersal::BlobInfo;
-use nomos_blend_message::{
-    crypto::{
-        keys::Ed25519PrivateKey,
-        proofs::{
-            quota::{inputs::prove::PublicInputs, ProofOfQuota},
-            selection::{inputs::VerifyInputs, ProofOfSelection},
-        },
-    },
-    encap::ProofsVerifier,
-};
-use nomos_blend_scheduling::message_blend::{BlendLayerProof, ProofsGenerator, SessionInfo};
 pub use nomos_blend_service::{
     core::{
         backends::libp2p::Libp2pBlendBackend as BlendBackend,
@@ -27,12 +13,12 @@ pub use nomos_blend_service::{
     },
     membership::service::Adapter as BlendMembershipAdapter,
 };
+use nomos_core::mantle::SignedMantleTx;
 pub use nomos_core::{
     codec,
     header::HeaderId,
     mantle::{select::FillSize as FillSizeWithTx, Transaction},
 };
-use nomos_core::{crypto::ZkHash, mantle::SignedMantleTx};
 pub use nomos_da_network_service::backends::libp2p::validator::DaNetworkValidatorBackend;
 use nomos_da_network_service::{
     api::http::HttApiAdapter, membership::handler::DaMembershipHandler, DaAddressbook,
@@ -69,6 +55,7 @@ pub use crate::config::{Config, CryptarchiaArgs, HttpArgs, LogArgs, NetworkArgs}
 use crate::{
     api::backend::AxumBackend,
     generic_services::{
+        blend::{BlendProofsGenerator, BlendProofsVerifier},
         DaMembershipAdapter, DaMembershipStorageGeneric, MembershipService, SdpService,
     },
 };
@@ -88,91 +75,9 @@ pub(crate) type TracingService = Tracing<RuntimeServiceId>;
 
 pub(crate) type NetworkService = nomos_network::NetworkService<NetworkBackend, RuntimeServiceId>;
 
-// TODO: Replace this with the actual verifier once the verification inputs are
-// successfully fetched by the Blend service.
-#[derive(Clone)]
-pub struct MockProofsVerifier;
-
-impl ProofsVerifier for MockProofsVerifier {
-    type Error = Infallible;
-
-    fn new() -> Self {
-        Self
-    }
-
-    fn verify_proof_of_quota(
-        &self,
-        _proof: ProofOfQuota,
-        _inputs: &PublicInputs,
-    ) -> Result<ZkHash, Self::Error> {
-        use groth16::Field as _;
-
-        Ok(ZkHash::ZERO)
-    }
-
-    fn verify_proof_of_selection(
-        &self,
-        _proof: ProofOfSelection,
-        _inputs: &VerifyInputs,
-    ) -> Result<(), Self::Error> {
-        Ok(())
-    }
-}
-
-pub struct MockProofsGenerator;
-
-#[async_trait]
-impl ProofsGenerator for MockProofsGenerator {
-    fn new(_session_info: SessionInfo) -> Self {
-        Self
-    }
-
-    /// Get or generate the next core `PoQ`, if the maximum allowance has not
-    /// been reached.
-    async fn get_next_core_proof(&mut self) -> Option<BlendLayerProof> {
-        Some(BlendLayerProof {
-            ephemeral_signing_key: Ed25519PrivateKey::generate(),
-            proof_of_quota: ProofOfQuota::from_bytes_unchecked([0; _]),
-            proof_of_selection: ProofOfSelection::from_bytes_unchecked([0; _]),
-        })
-    }
-    /// Get or generate the next leadership `PoQ`, if the maximum allowance has
-    /// not been reached.
-    async fn get_next_leadership_proof(&mut self) -> Option<BlendLayerProof> {
-        Some(BlendLayerProof {
-            ephemeral_signing_key: Ed25519PrivateKey::generate(),
-            proof_of_quota: ProofOfQuota::from_bytes_unchecked([0; _]),
-            proof_of_selection: ProofOfSelection::from_bytes_unchecked([0; _]),
-        })
-    }
-}
-
-pub type BlendProofsGenerator = MockProofsGenerator;
-pub type BlendProofsVerifier = MockProofsVerifier;
-
-pub(crate) type BlendCoreService = nomos_blend_service::core::BlendService<
-    BlendBackend,
-    PeerId,
-    BlendNetworkAdapter<RuntimeServiceId>,
-    BlendMembershipAdapter<MembershipService<RuntimeServiceId>, PeerId>,
-    BlendProofsGenerator,
-    BlendProofsVerifier,
-    RuntimeServiceId,
->;
-
-pub(crate) type BlendEdgeService = nomos_blend_service::edge::BlendService<
-    nomos_blend_service::edge::backends::libp2p::Libp2pBlendBackend,
-    PeerId,
-    <BlendNetworkAdapter<RuntimeServiceId> as nomos_blend_service::core::network::NetworkAdapter<
-        RuntimeServiceId,
-    >>::BroadcastSettings,
-    BlendMembershipAdapter<MembershipService<RuntimeServiceId>, PeerId>,
-    BlendProofsGenerator,
-    RuntimeServiceId,
->;
-
-pub(crate) type BlendService =
-    nomos_blend_service::BlendService<BlendCoreService, BlendEdgeService, RuntimeServiceId>;
+pub(crate) type BlendCoreService = generic_services::blend::BlendCoreService<RuntimeServiceId>;
+pub(crate) type BlendEdgeService = generic_services::blend::BlendEdgeService<RuntimeServiceId>;
+pub(crate) type BlendService = generic_services::blend::BlendService<RuntimeServiceId>;
 
 pub(crate) type BlockBroadcastService = broadcast_service::BlockBroadcastService<RuntimeServiceId>;
 
@@ -276,6 +181,8 @@ pub(crate) type ApiService = nomos_api::ApiService<
         NtpTimeBackend,
         DaNetworkApiAdapter,
         ApiStorageAdapter<RuntimeServiceId>,
+        BlendProofsGenerator,
+        BlendProofsVerifier,
         MB16,
     >,
     RuntimeServiceId,
